refactor: 디렉토리 구조 개선 및 커스텀훅 분리
