+++ src/components/schedule/index.ts
@@ -1,3 +1,4 @@
-export * from './ScheduleTable';
 export * from './ScheduleTables';
-export * from './ScheduleDndProvider';
+export * from './ScheduleTable';
+export * from './ScheduleWrapper';
+export * from './ScheduleHeader';
