+++ src/components/search/index.ts
@@ -1,3 +1,3 @@
-export * from './SearchDialog';
 export * from './SearchFilters';
+export * from './SearchDialog';
 export * from './LectureTable';
