+++ src/constants.ts
@@ -1,4 +1,4 @@
-export const DAY_LABELS = ["월", "화", "수", "목", "금", "토"] as const;
+export const DAY_LABELS = ['월', '화', '수', '목', '금', '토'] as const;
 
 export const CellSize = {
   WIDTH: 80,
@@ -7,3 +7,5 @@ export const CellSize = {
 
 export const 초 = 1000;
 export const 분 = 60 * 초;
+
+export const PAGE_SIZE = 100;
