+++ src/contexts/index.ts
@@ -1,1 +1,2 @@
 export * from './ScheduleContext';
+export * from './ScheduleDndProvider';
