+++ src/hooks/useSearchOptions.ts
@@ -1,0 +1,54 @@
+import { useState, useMemo } from 'react';
+import { SearchOption } from '../types';
+import { Lecture } from '../types';
+import { parseSchedule } from '../utils.ts';
+import { PAGE_SIZE } from '../constants.ts';
+
+export function useSearchOptions(lectures: Lecture[]) {
+  const [searchOptions, setSearchOptions] = useState<SearchOption>({
+    query: '',
+    grades: [],
+    days: [],
+    times: [],
+    majors: [],
+  });
+
+  const filteredLectures = useMemo(() => {
+    const { query = '', credits, grades, days, times, majors } = searchOptions;
+    return lectures
+      .filter(
+        (lecture) =>
+          lecture.title.toLowerCase().includes(query.toLowerCase()) ||
+          lecture.id.toLowerCase().includes(query.toLowerCase())
+      )
+      .filter((lecture) => grades.length === 0 || grades.includes(lecture.grade))
+      .filter((lecture) => majors.length === 0 || majors.includes(lecture.major))
+      .filter((lecture) => !credits || lecture.credits.startsWith(String(credits)))
+      .filter((lecture) => {
+        if (days.length === 0) {
+          return true;
+        }
+        const schedules = lecture.schedule ? parseSchedule(lecture.schedule) : [];
+        return schedules.some((s) => days.includes(s.day));
+      })
+      .filter((lecture) => {
+        if (times.length === 0) {
+          return true;
+        }
+        const schedules = lecture.schedule ? parseSchedule(lecture.schedule) : [];
+        return schedules.some((s) => s.range.some((time) => times.includes(time)));
+      });
+  }, [lectures, searchOptions]);
+
+  const lastPage = useMemo(
+    () => Math.ceil(filteredLectures.length / PAGE_SIZE),
+    [filteredLectures]
+  );
+
+  return {
+    searchOptions,
+    setSearchOptions,
+    filteredLectures,
+    lastPage,
+  };
+}
