+++ vite.config.ts
@@ -2,8 +2,11 @@ import { defineConfig as defineTestConfig, mergeConfig } from 'vitest/config';
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react-swc';
 
+const base = process.env.NODE_ENV === 'production' ? '/front_6th_chapter4-2/' : '';
+
 export default mergeConfig(
   defineConfig({
+    base,
     plugins: [react()],
   }),
   defineTestConfig({
@@ -12,9 +15,9 @@ export default mergeConfig(
       environment: 'jsdom',
       setupFiles: './src/setupTests.ts',
       coverage: {
-        reportsDirectory: "./.coverage",
-        reporter: ['lcov', 'json', 'json-summary']
+        reportsDirectory: './.coverage',
+        reporter: ['lcov', 'json', 'json-summary'],
       },
     },
   })
-)
+);
